:sparkles:  Feat: Category component 분리, Link 추가
