+++ components/index.ts
@@ -1,1 +1,2 @@
 export { default as CategoryList } from './CategoryList/CategoryList';
+export { default as Category } from './Category/Category';
